+++ protected/humhub/messages/bg/base.php
@@ -39,6 +39,7 @@ return [
     'Search' => 'Търси',
     '<strong>Confirm</strong> Action' => '',
     '<strong>Mail</strong> summary' => '',
+    'Actions' => '',
     'Add:' => '',
     'Allow' => '',
     'An error occurred while handling your last action. (Handler not found).' => '',