+++ protected/humhub/messages/ko/base.php
@@ -41,6 +41,7 @@ return [
     'Search' => '검색',
     '<strong>Confirm</strong> Action' => '',
     '<strong>Mail</strong> summary' => '',
+    'Actions' => '',
     'Add:' => '',
     'Allow' => '',
     'An error occurred while handling your last action. (Handler not found).' => '',