+++ protected/humhub/messages/ro/base.php
@@ -57,6 +57,7 @@ return [
     'You are not allowed to run this action.' => 'Nu aveți voie să executați această acțiune.',
     '<strong>Confirm</strong> Action' => '',
     '<strong>Mail</strong> summary' => '',
+    'Actions' => '',
     'An error occurred while handling your last action. (Handler not found).' => '',
     'An unexpected error occurred while loading the search result.' => '',
     'An unexpected error occurred. If this keeps happening, please contact a site administrator.' => '',