+++ protected/humhub/modules/user/messages/ro/notification.php
@@ -4,7 +4,7 @@ return array (
   'Mentionings' => '',
   'Receive Notifications when someone is following you.' => '',
   'Receive Notifications when someone mentioned you in a post.' => '',
-  'View Online' => '',
+  'View Online' => 'Vizualizați online',
   '{displayName} is now following you.' => '',
   '{displayName} just mentioned you in {contentTitle} "{preview}"' => '',
   '{displayName} mentioned you in {contentTitle}.' => '',